I am currently converting the labs overall AI process from MATLAB to python.
This commit adds the full_pipe.py file, which contains the full process for training a model on a set of images and masks. It also contains a function for extracting radiomic features from the images and masks.Th Please enter the commit message for your changes. Lines starting
It does not contain the nfold methods we traditionally use at this time nor does it contain the full compliment of radiomic features we have used. These will be added in the next version